panier2

+++ app/src/main/java/com/example/burgerking/DBHelper.java
@@ -6,23 +6,33 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class DBHelper extends SQLiteOpenHelper {
 
     public static final String DBname = "Login.db";
     public DBHelper(Context context) {
-        super(context,"Login.db", null, 1);
+        super(context,"Login.db", null,2);
     }
 
     @Override
     public void onCreate(SQLiteDatabase DB) {
         String queryString= "create Table user(id INTEGER PRIMARY KEY AUTOINCREMENT,username TEXT,email TEXT,phone INTEGER,password TEXT)";
         DB.execSQL(queryString);
-        //DB.execSQL("create table shoopingCart(name TEXT,price TEXT,image BLOB)");
+        DB.execSQL("create table shoopingCart(name TEXT ,price TEXT ,image BLOB)");
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase DB, int i, int i1) {
-        DB.execSQL("drop table if exists user");
+    public void onUpgrade(SQLiteDatabase DB, int oldVersion, int newVersion) {
+        if(newVersion > oldVersion){
+            DB.execSQL("drop table if exists user");
+            DB.execSQL("drop table if exists shoopingCart");
+            onCreate(DB);
+        }
+
+
+
 
     }
     public Boolean insertData(String username,String email,int phone,String password){
@@ -54,12 +64,22 @@ public class DBHelper extends SQLiteOpenHelper {
         if(cursor.getCount() > 0) return true;
         else return false;
     }
-   /* public void AddshoopingCart(String name,String price,byte[] image){
+
+
+
+   public void AddshoopingCart(String name,String price,byte[] image){
         SQLiteDatabase db = getWritableDatabase();
         ContentValues data= new ContentValues();
         data.put("name",name);
         data.put("price",price);
         data.put("image",image);
         db.insert("shoopingCart",null,data);
+    }
+   /* public int countShoopingcart(){
+        SQLiteDatabase db = getReadableDatabase();
+
+        Cursor cn = db.rawQuery("select * from shoopingCart",null);
+        cn.moveToFirst();
+        return cn.getCount();
     }*/
 }
